web-animations/timing-model/animations/playing-an-animation.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Tests for current time</title>
<link rel="help"
      href="https://w3c.github.io/web-animations/#playing-an-animation-section">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

test(function(t) {
  var animation = createDiv(t).animate(null, 100 * MS_PER_SEC);
  animation.currentTime = 1 * MS_PER_SEC;
  assert_times_equal(animation.currentTime, 1 * MS_PER_SEC);
  animation.play();
  assert_times_equal(animation.currentTime, 1 * MS_PER_SEC);
}, 'play() leaves currentTime unchanged on running animation');

test(function(t) {
  var animation = createDiv(t).animate(null, 100 * MS_PER_SEC);
  animation.finish();
  assert_times_equal(animation.currentTime, 100 * MS_PER_SEC);
  animation.play();
  assert_times_equal(animation.currentTime, 0);
}, 'play() seeks back to start on finished animation');

test(function(t) {
  var animation = createDiv(t).animate(null, 100 * MS_PER_SEC);
  animation.playbackRate = -1;
  animation.currentTime = 0;
  assert_times_equal(animation.currentTime, 0);
  animation.play();
  assert_times_equal(animation.currentTime, 100 * MS_PER_SEC);
}, 'play() seeks to end on reverse finished animation');

</script>
</body>
